refactor(chains): extract categorization chain builder
